+++ test/grouping.jl
@@ -2918,16 +2918,19 @@ end
             elseif cutoff == 10
                 @test isempty(gdf2)
             end
+            @test filter(predicate, gdf1, ungroup=true) == DataFrame(gdf2)
         end
 
-        @test_throws TypeError filter(x -> 1, groupby_checked(df, :g1))
-        @test_throws TypeError filter(r"x" => (x...) -> 1, groupby_checked(df, :g1))
-        @test_throws TypeError filter(AsTable(r"x") => (x...) -> 1, groupby_checked(df, :g1))
+        for ug in (true, false)
+            @test_throws TypeError filter(x -> 1, groupby_checked(df, :g1), ungroup=ug)
+            @test_throws TypeError filter(r"x" => (x...) -> 1, groupby_checked(df, :g1), ungroup=ug)
+            @test_throws TypeError filter(AsTable(r"x") => (x...) -> 1, groupby_checked(df, :g1), ungroup=ug)
 
-        @test_throws ArgumentError filter(r"y" => (x...) -> true, groupby_checked(df, :g1))
-        @test_throws ArgumentError filter([] => (x...) -> true, groupby_checked(df, :g1))
-        @test_throws ArgumentError filter(AsTable(r"y") => (x...) -> true, groupby_checked(df, :g1))
-        @test_throws ArgumentError filter(AsTable([]) => (x...) -> true, groupby_checked(df, :g1))
+            @test_throws ArgumentError filter(r"y" => (x...) -> true, groupby_checked(df, :g1), ungroup=ug)
+            @test_throws ArgumentError filter([] => (x...) -> true, groupby_checked(df, :g1), ungroup=ug)
+            @test_throws ArgumentError filter(AsTable(r"y") => (x...) -> true, groupby_checked(df, :g1), ungroup=ug)
+            @test_throws ArgumentError filter(AsTable([]) => (x...) -> true, groupby_checked(df, :g1), ungroup=ug)
+        end
     end
 end